# packages/api-types/openapi/v1.yaml
openapi: 3.0.3
info:
  title: EmpOps API
  version: 1.0.0
  description: Step 2 — Teams, hierarchy, and org structure (Laravel and Go parity).
servers:
  - url: http://localhost:8000/api/v1
    description: Laravel
  - url: http://localhost:8080/api/v1
    description: Go
paths:
...
  /companies/{companyId}/teams:
    get:
      summary: List teams
      operationId: listTeams
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      responses:
        "200":
          description: Teams
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Team"
    post:
      summary: Create team
      operationId: createTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Team"
  /companies/{companyId}/teams/{teamId}:
    get:
      summary: Get team
      operationId: getTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Team"
    patch:
      summary: Update team
      operationId: updateTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/TeamId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamRequest"
      responses:
        "200":
          description: Updated
    delete:
      summary: Delete team
      operationId: deleteTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/TeamId"
      responses:
        "200":
          description: Deleted
...
  /companies/{companyId}/employees/{employeeId}/managers:
    get:
      summary: List managers for employee
      operationId: listManagers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/EmployeeId"
      responses:
        "200":
          description: Managers
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/EmployeeSummary"
    post:
      summary: Assign manager
      operationId: assignManager
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/EmployeeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignManagerRequest"
      responses:
        "201":
          description: Manager assigned
  /companies/{companyId}/employees/{employeeId}/managers/{managerId}:
    delete:
      summary: Unassign manager
      operationId: unassignManager
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/EmployeeId"
        - $ref: "#/components/parameters/ManagerId"
      responses:
        "200":
          description: Manager unassigned
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CompanyId:
      name: companyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EmployeeId:
      name: employeeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    TeamId:
      name: teamId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ManagerId:
      name: managerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    ApiEnvelope:
      type: object
      required: [success, message, data, error, timestamp]
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
        error:
          nullable: true
        timestamp:
          type: string
          format: date-time
...
    EmployeeSummary:
      type: object
      required: [id, first_name, last_name, email]
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
...
    Team:
      type: object
      required: [id, company_id, name, members, member_count]
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        leader:
          $ref: "#/components/schemas/EmployeeSummary"
          nullable: true
        members:
          type: array
          items:
            $ref: "#/components/schemas/EmployeeSummary"
        member_count:
          type: integer
    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
    UpdateTeamRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
...
    AssignManagerRequest:
      type: object
      required: [manager_id]
      properties:
        manager_id:
          type: string
          format: uuid
...